# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: SmartSME API
  description: |
    SmartSME Backend API with Role-Based Access Control (RBAC)
    
    ## Authentication
    - JWT Bearer token authentication
    - Guest access (no authentication) with read-only permissions
    
    ## Roles
    - **OWNER**: Full access to all resources
    - **ADMIN**: Administrative access
    - **PLANT_HEAD**: Plant management access
    - **SHIFT_INCHARGE**: Production and secondary process management
    - **STORES_INCHARGE**: Inventory and dispatch management
    - **PRODUCTION_EMPLOYEE**: Production operations
    - **SECONDARY_PROCESS_EMPLOYEE**: Secondary process operations
    - **ACCOUNTANT**: Billing and accounting access
    - **GUEST**: Read-only access (no authentication required)
  version: 1.0.0
  contact:
    name: SmartSME Team
    email: support@smartsme.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://127.0.0.1:8080
    description: Development server
  - url: https://staging-api.smartsme.com
    description: Staging server
  - url: https://api.smartsme.com
    description: Production server

security:
  - BearerAuth: []
  - {}  # Allow guest access

paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                  description: Username, email, or mobile number
                  example: "tamil"
                password:
                  type: string
                  description: User password
                  example: "password123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  accessToken:
                    type: string
                    description: JWT access token
                  refreshToken:
                    type: string
                    description: JWT refresh token
        '401':
          description: Invalid credentials

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    description: New JWT access token
        '401':
          description: Invalid or expired refresh token

  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user information
      description: Retrieve authenticated user's profile and roles
      responses:
        '200':
          description: User information retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Authentication required
        '404':
          description: User not found

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout current user (client-side token removal)
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout successful"

  /permissions/check:
    post:
      tags:
        - Permissions
      summary: Check user permissions
      description: |
        Check if current user (or guest) has specific permissions.
        Guests only have read permissions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - permissions
              properties:
                permissions:
                  type: array
                  items:
                    type: string
                  example: ["PRODUCTION_READ", "PRODUCTION_CREATE"]
      responses:
        '200':
          description: Permission check results
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasPermissions:
                    type: object
                    additionalProperties:
                      type: boolean
                    example:
                      PRODUCTION_READ: true
                      PRODUCTION_CREATE: false
        '400':
          description: Invalid request format

  /permissions/me:
    get:
      tags:
        - Permissions
      summary: Get user permissions
      description: |
        Get all permissions for current user or guest.
        Guests receive only read permissions.
      responses:
        '200':
          description: User permissions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Permission'

  /actions/check:
    post:
      tags:
        - Actions
      summary: Check user actions
      description: |
        Check if current user (or guest) can perform specific actions.
        Guests can only perform READ actions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - actions
              properties:
                actions:
                  type: array
                  items:
                    type: string
                  example: ["READ", "CREATE", "UPDATE"]
      responses:
        '200':
          description: Action check results
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasActions:
                    type: object
                    additionalProperties:
                      type: boolean
                    example:
                      READ: true
                      CREATE: false
                      UPDATE: false
        '400':
          description: Invalid request format

  /actions/me:
    get:
      tags:
        - Actions
      summary: Get user actions
      description: |
        Get all actions available to current user or guest.
        Guests receive only READ action.
      responses:
        '200':
          description: User actions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  actions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Action'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        username:
          type: string
          description: Unique username
          example: "tamil"
        firstName:
          type: string
          description: First name
          example: "Tamilselvan"
        lastName:
          type: string
          description: Last name
          example: "M"
        name:
          type: string
          description: Full name
          example: "Tamilselvan M"
        email:
          type: string
          format: email
          description: Email address
          example: "tamil@test.com"
        mobile:
          type: string
          description: Mobile number
          example: "7889043243"
        address:
          type: string
          description: Address
          example: "Medavakkam, Chennai"
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'

    Role:
      type: object
      properties:
        id:
          type: integer
          description: Role ID
          example: 1
        name:
          type: string
          description: Role name
          example: "OWNER"

    Permission:
      type: object
      properties:
        id:
          type: integer
          description: Permission ID
          example: 1
        name:
          type: string
          description: Permission name
          example: "PRODUCTION_CREATE"
        resource:
          type: string
          description: Resource type
          example: "PRODUCTION"

    Action:
      type: object
      properties:
        id:
          type: integer
          description: Action ID
          example: 1
        name:
          type: string
          description: Action name
          example: "CREATE"
        description:
          type: string
          description: Action description
          example: "Create new records"

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Authentication required"

tags:
  - name: Authentication
    description: User authentication and session management
  - name: Permissions
    description: Role-based permission checking
  - name: Actions
    description: Role-based action checking